+++ app/src/utils/eventUtils.ts
@@ -5,85 +5,119 @@ import { RRule } from '@heinlein-video/rrule';
 import {
   Event,
   EventException,
-  EventType,
   InviteStatus,
   isEvent,
+  isRecurringEvent,
   isTimelessEvent,
   RecurringEvent,
 } from '@opentalk/rest-api-rtk-query';
-import { addMonths, isAfter, subMonths } from 'date-fns';
-import _ from 'lodash';
+import { addMonths, subMonths } from 'date-fns';
+import { cloneDeep, orderBy } from 'lodash';
 
 import { TimePerspectiveFilter } from '../pages/Dashboard/EventsOverviewPage/EventsOverviewPage';
+import { getISOStringWithoutMilliseconds } from './timeUtils';
 
-export const getExpandedEvents = (
+const DEFAULT_MONTHS_CONSIDERED = 3;
+
+const mapDateToRecurringEvent = (recurrenceDate: Date, initialEvent: RecurringEvent) => {
+  const recurringEvent = cloneDeep(initialEvent);
+  const recurrenceStartDate = new Date(initialEvent.startsAt.datetime);
+  const recurrenceEndDate = new Date(initialEvent.endsAt.datetime);
+
+  const startDate = new Date(recurrenceStartDate);
+  startDate.setDate(recurrenceDate.getDate());
+  startDate.setMonth(recurrenceDate.getMonth());
+  startDate.setUTCFullYear(recurrenceDate.getUTCFullYear());
+  recurringEvent.startsAt = {
+    datetime: getISOStringWithoutMilliseconds(startDate),
+    timezone: initialEvent.startsAt.timezone,
+  };
+
+  const endDate = new Date(recurrenceEndDate);
+  endDate.setDate(recurrenceDate.getDate());
+  endDate.setMonth(recurrenceDate.getMonth());
+  endDate.setUTCFullYear(recurrenceDate.getUTCFullYear());
+  recurringEvent.endsAt = {
+    datetime: getISOStringWithoutMilliseconds(endDate),
+    timezone: initialEvent.endsAt.timezone,
+  };
+
+  return recurringEvent;
+};
+
+const createRecurrenceEventInstances = (
+  event: RecurringEvent,
+  maxMonths: number = DEFAULT_MONTHS_CONSIDERED,
+  filter?: TimePerspectiveFilter
+) => {
+  const today = new Date();
+  const recurrenceStartDate = new Date(event.startsAt.datetime);
+
+  let windowStartOffset = 0;
+  let windowEndOffset = maxMonths;
+
+  if (filter === TimePerspectiveFilter.Past) {
+    windowStartOffset = maxMonths;
+    windowEndOffset = 0;
+  }
+
+  const windowStartDate = subMonths(today, windowStartOffset);
+  const windowEndDate = addMonths(today, windowEndOffset);
+
+  const partialRule = RRule.parseString(event.recurrencePattern[0]);
+
+  partialRule.dtstart = recurrenceStartDate;
+  partialRule.until = windowEndDate;
+
+  const rule = new RRule({
+    ...partialRule,
+  });
+
+  const generatedRecurrenceDates = rule.between(windowStartDate, windowEndDate, true);
+
+  return generatedRecurrenceDates.map((generatedRecurrenceDate) =>
+    mapDateToRecurringEvent(generatedRecurrenceDate, event)
+  );
+};
+
+export enum SortDirection {
+  ASC = 'asc',
+  DESC = 'desc',
+}
+
+export const orderEventsByDate = (events: Event[], sortDirection: SortDirection = SortDirection.ASC) =>
+  orderBy(
+    events,
+    [
+      (event: Event) => {
+        if (isTimelessEvent(event)) {
+          return new Date(event.createdAt);
+        }
+        return new Date(event.startsAt.datetime);
+      },
+    ],
+    [sortDirection]
+  );
+
+export const appendRecurrenceEventInstances = (
   eventList: (EventException | Event)[],
   filterDeclined?: boolean,
-  maxEntries?: number,
-  firstDateIsoString?: string,
   maxMonths?: number,
   filter?: TimePerspectiveFilter
 ): Event[] => {
   const events = Array<Event>();
-  const DEFAULT_MONTHS_CONSIDERED = 3;
-  eventList.forEach((event) => {
-    if (isEvent(event) && (filterDeclined ? event.inviteStatus !== InviteStatus.Declined : true)) {
-      if (!isTimelessEvent(event) && event.type === EventType.Recurring) {
-        const startDate = new Date(event.startsAt.datetime);
-        const endDate = new Date(event.endsAt.datetime);
-        const startHour = startDate.getHours();
-        const startMin = startDate.getMinutes();
-        const duration = endDate.getTime() - startDate.getTime();
-        const recurrencePattern = event.recurrencePattern[0];
 
-        const rule = RRule.fromString(
-          `DTSTART:${event.startsAt.datetime.replaceAll(':', '').replaceAll('-', '')};
-              ${recurrencePattern}`
-        );
-
-        const maxConsideredMonths = maxMonths ? maxMonths : DEFAULT_MONTHS_CONSIDERED;
-
-        // If timePerspective is in the future, start date is today and end date is ${maxConsideredMonths} from today
-        // if timePerspective is in the past, start date is calculated by substracting ${maxConsideredMonths} from today and adding today as end date
-        const recurrenceInterval =
-          filter === TimePerspectiveFilter.Future
-            ? { start: new Date(), end: addMonths(new Date(), maxConsideredMonths) }
-            : { start: subMonths(new Date(), maxConsideredMonths), end: startDate };
-
-        const timeline = rule.between(recurrenceInterval.start, recurrenceInterval.end, true);
-
-        const firstIsoDate = firstDateIsoString ? new Date(firstDateIsoString) : undefined;
-        const firstDateTime = filter === TimePerspectiveFilter.Future ? new Date() : firstIsoDate;
-
-        timeline
-          .slice(0, maxEntries)
-          .filter((value) => firstDateTime === undefined || isAfter(new Date(value.toISOString()), firstDateTime))
-          .forEach((startDate) => {
-            const startDateAdjusted = startDate;
-            startDateAdjusted.setHours(startHour);
-            startDateAdjusted.setMinutes(startMin);
-            const recurringEvent = { ...event } as RecurringEvent;
-            const endDate = new Date(startDateAdjusted.getTime() + duration);
-            recurringEvent.startsAt = { datetime: startDateAdjusted.toISOString(), timezone: event.startsAt.timezone };
-            recurringEvent.endsAt = { datetime: endDate.toISOString(), timezone: event.endsAt.timezone };
-            events.push(recurringEvent);
-          });
+  eventList
+    .filter((event): event is Event => isEvent(event))
+    .filter((event) => (filterDeclined ? event.inviteStatus !== InviteStatus.Declined : true))
+    .forEach((event) => {
+      if (!isTimelessEvent(event) && isRecurringEvent(event)) {
+        createRecurrenceEventInstances(event, maxMonths, filter).forEach((event) => {
+          events.push(event);
+        });
       } else {
         events.push(event);
       }
-    }
-  });
-
-  const sortedEvents = _.sortBy(events, function (event: Event) {
-    if (isTimelessEvent(event)) {
-      return new Date(event.createdAt).getTime();
-    }
-    return new Date(event.startsAt.datetime).getTime();
-  });
-
-  if (maxEntries) {
-    return sortedEvents.slice(0, maxEntries);
-  }
-
-  return sortedEvents;
+    });
+  return events;
 };